RRSet: fix serialization for bools and structs (#31)
For booleans which default to true, it's important not to omitempty.

For TCPoolProfile.BackupRecord, we want to omit it if its RData is
empty, and the easiest way to accomplish this is to stack omitempty for
both. This means we need to be cautious not to set a FailoverDelay
without any RData.